requirements.txt
python-dotenv
stripe
pandas